unburden by preserving docker layers

+++ minor-upgrade.Dockerfile
@@ -1,14 +1,16 @@
-FROM golang:1.23 AS builder
+FROM golang:1.23.6 AS builder
 WORKDIR /app
 
 COPY . ./
 RUN go mod download
 RUN go build -C ./cmd/minor-upgrade -o main
 
-FROM ubuntu:latest
-RUN set -x && apt-get update && apt-get install -y \
-    ca-certificates && \
-    rm -rf /var/lib/apt/lists/*
+FROM debian:bookworm-slim
+RUN --mount=target=/var/lib/apt/lists,type=cache,sharing=locked \
+    --mount=target=/var/cache/apt/,type=cache,sharing=locked \
+    set -x && apt-get update && apt-get install -y \
+    ca-certificates && rm -rf /var/lib/apt/lists/*
+
 COPY --from=builder /app/cmd/minor-upgrade/main /app/server
 WORKDIR /app
 CMD ["/app/server"]
